Handle audio amp level event, based on patch from Reinhard Nissl.
Corrected str_type_return arg for XrmGetResource(): It's a pointer address,
not a pointer array (* and [] can be ambiguous in C); the array[20] is sometimes
found in open sources, seems it stems from the X windows description which says
"a series of from 1 to 20 probes" is made in the resource database.

Needless memsets removed, adjusted filename buffer sizes, text corrections.