Cleaner kamikaze in event processing (access to the freed memory when callback destroys own object).
Events in xitk are passed via callbacks to widgets according to registered widget lists. During the callbacks it should be safe to change widget list, but before all callbacks ends in the one event, it should be preserved xitk_widget_list_t structure itself. So it's only marked for pending removal on locked "fx".