Add Close command
The Close command is similar to the Stop command as it stops playing the
active stream and additionally closes the xine stream. This forces the
input plugin to close their files and possible network connections.

I used the Close command in connection with the suspend mode on a diskless
network client. You do not have to quit the player when going to suspend
mode and on resume start playing the network stream is faster then loading
the player from scratch.