Fix deadlock while switching to/from fullscreen mode.
When the video output driver (e. g. vdpau) has no need to use XLockDisplay()
to synchronize its output operation, it is likely that we get an output
callback while we are about to change the drawable and hence deadlock.
Using a timed lock avoids the deadlock and the callback returns immediately.
As we are about to change drawable anyway, leaving the callback early
doesn't matter at that time.