Fix volume 0 at start Right after startup, xine-ui remembers the volume in 'original_level'. Just before closing, xine-ui restores the volume from 'original_level'.
In pulseaudio, the volume can only be inquired if a stream is open. Without
it, the volume will always be 0. As 'original_level' is taken right at the
start when no stream is open, the restore operation sets the volume always
to 0. Pulseaudio remembers this value and brings it up again at the next
startup.